FactForge/backend/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
httpx==0.27.0
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1; platform_system != "Windows"
numpy==1.26.4
pytest==8.2.0
